Added FlightController search endpoint, ScheduledJobService, Basic Authentication and Swagger UI

+++ src/main/java/com/cenktu/flightsearch/controller/AirportController.java
@@ -2,6 +2,8 @@ package com.cenktu.flightsearch.controller;
 
 import com.cenktu.flightsearch.model.Airport;
 import com.cenktu.flightsearch.service.AirportService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,33 +16,43 @@ public class AirportController {
     private AirportService airportService;
 
     public AirportController(AirportService airportService){
+
         this.airportService=airportService;
     }
 
     @GetMapping
-    public List<Airport> getAllAirports(){
-        return airportService.getAllAirports();
+    public ResponseEntity<List<Airport>> getAllAirports(){
+        return ResponseEntity.ok(airportService.getAllAirports());
     }
 
     @GetMapping("/{airportId}")
-    public Optional<Airport> getSingleAirport(@PathVariable Long airportId){
-        return airportService.getSingleAirport(airportId);
+    public ResponseEntity<Airport> getSingleAirport(@PathVariable Long airportId){
+        return ResponseEntity.ok(airportService.getSingleAirport(airportId));
     }
 
 
     @PostMapping
-    public Airport createAirport(@RequestBody Airport newAirport){
-        return airportService.createAirport(newAirport);
+    public ResponseEntity<Airport> createAirport(@RequestBody Airport newAirport){
+        return ResponseEntity.ok(airportService.createAirport(newAirport));
     }
 
     @PutMapping("/{airportId}")
-    public Airport updateAirport(@PathVariable Long airportId,@RequestBody Airport updatedAirport){
-        return airportService.updateAirport(airportId,updatedAirport);
+    public ResponseEntity<Airport> updateAirport(@PathVariable Long airportId,@RequestBody Airport updatedAirport){
+        Airport optionalAirport = airportService.getSingleAirport(airportId);
+        if(optionalAirport!=null){
+            return ResponseEntity.ok(airportService.updateAirport(airportId,updatedAirport));
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{airportId}")
-    public void deleteAirport(@PathVariable Long airportId){
-        airportService.deleteAirport(airportId);
+    public ResponseEntity<String> deleteAirport(@PathVariable Long airportId){
+        Airport optionalAirport = airportService.getSingleAirport(airportId);
+        if(optionalAirport!=null){
+            airportService.deleteAirport(airportId);
+            return ResponseEntity.ok("Airport deleted successfully");
+        }
+        return ResponseEntity.notFound().build();
     }
 
 }
