+++ src/main/java/com/cenktu/flightsearch/controller/FlightController.java
@@ -1,9 +1,15 @@
 package com.cenktu.flightsearch.controller;
 
+
 import com.cenktu.flightsearch.model.Flight;
+import com.cenktu.flightsearch.request.FlightRequest;
 import com.cenktu.flightsearch.service.FlightService;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,27 +22,54 @@ public class FlightController {
         this.flightService =flightService;
     }
     @GetMapping
-    public List<Flight> getAllFlights(){
-        return flightService.getAllFlights();
+    public ResponseEntity<List<Flight>> getAllFlights(){
+        return ResponseEntity.ok(flightService.getAllFlights());
+
     }
 
     @GetMapping ("/{flightId}")
-    public Optional<Flight> getSingleFlight(@PathVariable Long flightId){
-        return flightService.getSingleFlight(flightId);
+    public ResponseEntity<Flight> getSingleFlight(@PathVariable Long flightId){
+        Optional<Flight> optionalFlight = flightService.getSingleFlight(flightId);
+        if(optionalFlight.isPresent()){
+            Flight flight = optionalFlight.get();
+            return ResponseEntity.ok(flight);
+        }
+
+        return ResponseEntity.notFound().build();
+    }
+
+    @GetMapping("/search")
+    public ResponseEntity<List<Flight>> searchFlights(@RequestParam String departureCity,
+                                                      @RequestParam String arrivalCity,
+                                                      @RequestParam  @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDate departureDate,
+                                                      @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDate returnDate){
+        List<Flight> flights = flightService.searchFlights(departureCity, arrivalCity, departureDate, returnDate);
+        return ResponseEntity.ok(flights);
     }
 
     @PostMapping
-    public Flight createFlight(@RequestBody Flight newFlight){
-        return flightService.createFlight(newFlight);
+    public ResponseEntity<Flight> createFlight(@RequestBody FlightRequest newFlightRequest){
+        return ResponseEntity.ok(flightService.createFlight(newFlightRequest));
     }
 
     @PutMapping("/{flightId}")
-    public Flight updateFlight(@PathVariable Long flightId, @RequestBody Flight updatedFlight){
-        return flightService.updateFlight(flightId,updatedFlight);
+    public ResponseEntity<Flight> updateFlight(@PathVariable Long flightId, @RequestBody FlightRequest updatedFlightRequest){
+        Optional<Flight> optionalFlight = flightService.getSingleFlight(flightId);
+        if(optionalFlight.isPresent()){
+            return ResponseEntity.ok(flightService.updateFlight(flightId,updatedFlightRequest));
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{flightId}")
-    public void deleteFlight(@PathVariable Long flightId){
-        flightService.deleteFlight(flightId);
+    public ResponseEntity<String> deleteFlight(@PathVariable Long flightId){
+        Optional<Flight> optionalFlight = flightService.getSingleFlight(flightId);
+        if(optionalFlight.isPresent()){
+            flightService.deleteFlight(flightId);
+            return ResponseEntity.ok("Flight deleted successfully");
+        }
+        return ResponseEntity.notFound().build();
     }
+
+
 }
